+++ docs/docs/release_notes.md
@@ -2,7 +2,8 @@
 - Check ws redis connection on the `init`
 - Refactor `Monitoring` class and usage
 - Improve `logging` config
- 
+- Check database connection before query
+
 ### 3.1.3
 - Add `Image` base class 
 - Add `size` to `File` base class